refactor(ai): split tool schemas and formatters
Extract static tool definitions and response formatting helpers from the weather tool executor while preserving the ai_tools compatibility exports.

Tested: uv run pytest -q -n 0 --tb=short tests/test_ai_tools.py tests/test_ai_tools_extended.py tests/test_ai_tools_formatters.py; uv run ruff check src tests scripts; uv run pyright

Confidence: high
